Move MNIST tests to examples
